requirements.txt
anyio==4.12.1
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.3.1
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
inquirerpy==0.3.4
markdown-it-py==4.0.0
mdurl==0.1.2
notebooklm-py==0.3.4
pfzy==0.3.4
playwright==1.58.0
prompt_toolkit==3.0.52
pyee==13.0.1
Pygments==2.19.2
questionary==2.1.1
requests==2.32.5
rich==14.3.3
typing_extensions==4.15.0
urllib3==2.6.3
wcwidth==0.6.0
